Fixed tests with phantom * Executed `phantomjs` instead of `mocha-phantomjs` * Used `require('phantomjs').path` to get bin path * Used `mocha-phantomjs` as script for `phantomjs` * Moved runPhantom function to lib
