+++ test/techs/bh-client-module.test.js
@@ -1,9 +1,7 @@
 var fs = require('fs'),
     path = require('path'),
-    vow = require('vow'),
     mock = require('mock-fs'),
     promisify = require('vow-node').promisify,
-    exec = require('child_process').exec,
     http = require('http'),
     serveStatic = require('serve-static'),
     finalhandler = require('finalhandler'),
@@ -16,6 +14,7 @@ var fs = require('fs'),
     chaiFilename = require.resolve('chai/chai.js'),
     ymFilename = require.resolve('ym/modules.js'),
     writeFile = require('../lib/write-file'),
+    runPhantom = require('../lib/run-phantom'),
     serve, server, listen;
 
 describe('bh-client-module', function () {
@@ -202,7 +201,9 @@ describe('bh-client-module', function () {
 
                     return listen(3000);
                 })
-                .then(runPhantom)
+                .then(function () {
+                    return runPhantom('http://localhost:3000/index.html');
+                })
                 .fail(function (err) {
                     throw err;
                 });
@@ -243,7 +244,9 @@ describe('bh-client-module', function () {
 
                     return listen(3000);
                 })
-                .then(runPhantom)
+                .then(function () {
+                    return runPhantom('http://localhost:3000/index.html');
+                })
                 .fail(function (err) {
                     throw err;
                 });
@@ -281,7 +284,9 @@ describe('bh-client-module', function () {
 
                     return listen(3000);
                 })
-                .then(runPhantom)
+                .then(function () {
+                    return runPhantom('http://localhost:3000/index.html');
+                })
                 .fail(function (err) {
                     throw err;
                 });
@@ -335,7 +340,9 @@ function runTest(testContent, options, template, lib) {
 
             return listen(3000);
         })
-        .then(runPhantom)
+        .then(function () {
+            return runPhantom('http://localhost:3000/index.html');
+        })
         .then(function () {
             server.close();
         })
@@ -358,33 +365,3 @@ function generateTest(json, result) {
     ].join('\n');
 }
 
-function runPhantom() {
-    var defer = vow.defer();
-
-    exec(
-        'node_modules/mocha-phantomjs/bin/mocha-phantomjs --reporter json http://localhost:3000/index.html',
-        function (err, stdout) {
-            if (err) {
-                var json = JSON.parse(stdout),
-                    errors = json.tests.filter(function (test) {
-                        return test.err;
-                    }),
-                    testError = errors[0].err;
-
-                if (testError) {
-                    var stack = testError.stack;
-                    testError = new Error(testError.message);
-                    testError.stack = stack;
-                } else {
-                    testError = err;
-                }
-
-                defer.reject(testError);
-            } else {
-                defer.resolve();
-            }
-        }
-    );
-
-    return defer.promise();
-}
